requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
motor>=3.3.0
beanie>=1.20.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
pymongo[srv]>=4.0
